Add AxiosWrapper component for handling token expiration in axios interceptor
